Plugins Architecture Definition:
fenix-catalog-bridge is now meant as a plugin system. Components use a catalog plugin to delegate a certain set of operation to the bridge.

Plugin registration is automatic: each plugin will register itself within a global repository (a window variable) mapped with the unique id of the component it will serve.

The linking between Visualization Component (external to catalog but with a catalog plugin) and the catalog bridge is up to the controller. The linking is based on two different methodologies: 
-bridge will receive the two components (possibly they are the same one), retrieve the relative plugin from the global plugin repository and will accomplish to operation required executing operations eventually asking parameter to plugins and delegating a possible callback function to plugins.
- bridge will receive one component and a callback function, retrieve the component plugin from the global plugin repository and will accomplish to operation required executing operations eventually asking parameter to the plugin and invoking the callback passing the processed input.